cambios en entrada de parametros a long en Usuario GetById

+++ Decimatio.Domain/Interfaces/IUsuarioRepository.cs
@@ -4,9 +4,9 @@
     {
         Task<IEnumerable<Usuario>> GetAllUsers();
         Task<IEnumerable<Usuario>> GetAllUsersFilter(string filtro);
-        Task<Usuario> GetById(int idUsuario);
+        Task<Usuario> GetById(long idUsuario);
         Task AddUsuario(Usuario usuario);
         Task<bool> UpdateUsuario(Usuario usuario);
-        Task<bool> DeleteUsuario(int idUsuario);
+        Task<bool> DeleteUsuario(long idUsuario);
     }
 }
